refactor: extract EditServerModal component shared by Dashboard and ServerDetail
